+++ packages/next-common/components/summary/democracyAllMyDelegationPopup/index.js
@@ -6,7 +6,10 @@ const Popup = styled(PopupOrigin)`
   width: 480px;
 `;
 
-export default function MyDelegationPopup({ delegations, setShow = () => {} }) {
+export default function AllMyDelegationPopup({
+  delegations,
+  setShow = () => {},
+}) {
   return (
     <Popup title="My Delegation" onClose={() => setShow(false)}>
       <MyDelegationTabList delegations={delegations} />
